Modified app.js to return object instead of writing to file; moved code to write to file to cli.js
