+++ src/cli/app.js
@@ -129,19 +129,23 @@ async function mcodeApp(Client, fromDate, toDate, pathToConfig, pathToRunLogs, d
     }
   }
 
+
+
+
   // Finally, save the data to disk
-  const outputPath = './output';
-  if (!fs.existsSync(outputPath)) {
-    logger.info(`Creating directory ${outputPath}`);
-    fs.mkdirSync(outputPath);
-  }
-  // For each bundle in our extractedData, write it to our output directory
-  extractedData.forEach((bundle, i) => {
-    const outputFile = path.join(outputPath, `mcode-extraction-patient-${i + 1}.json`);
-    logger.debug(`Logging mCODE output to ${outputFile}`);
-    fs.writeFileSync(outputFile, JSON.stringify(bundle), 'utf8');
-  });
-  logger.info(`Successfully logged ${extractedData.length} mCODE bundle(s) to ${outputPath}`);
+  // const outputPath = './output';
+  // if (!fs.existsSync(outputPath)) {
+  //   logger.info(`Creating directory ${outputPath}`);
+  //   fs.mkdirSync(outputPath);
+  // }
+  // // For each bundle in our extractedData, write it to our output directory
+  // extractedData.forEach((bundle, i) => {
+  //   const outputFile = path.join(outputPath, `mcode-extraction-patient-${i + 1}.json`);
+  //   logger.debug(`Logging mCODE output to ${outputFile}`);
+  //   fs.writeFileSync(outputFile, JSON.stringify(bundle), 'utf8');
+  // });
+  // logger.info(`Successfully logged ${extractedData.length} mCODE bundle(s) to ${outputPath}`);
+  return extractedData;
 }
 
 module.exports = {
